# Conflicts:
#	server/player/handler.go

+++ server/player/playerdb/json.go
@@ -92,7 +92,7 @@ func (p *Provider) toJson(d player.Config, w *world.World) jsonData {
 			MainHandSlot: uint32(d.HeldSlot),
 		}),
 		EnderChestInventory: encodeItems(d.EnderChestInventory.Slots()),
-		Dimension:           uint8(dim),
+		Dimension:           int32(dim),
 	}
 }
 
@@ -115,7 +115,7 @@ type jsonData struct {
 	Effects                          []jsonEffect
 	FireTicks                        int64
 	FallDistance                     float64
-	Dimension                        uint8
+	Dimension                        int32
 }
 
 type jsonInventoryData struct {
